refactor: home screen에서 clipboard messenger 분리
